+++ setup.py
@@ -24,7 +24,7 @@ setup_keywords['provides'] = [setup_keywords['name']]
 setup_keywords['requires'] = ['Python (>3.7.0)']
 setup_keywords['install_requires'] = [
     'xarray', 'h5netcdf', 'torch', 'seaborn', 'smart-open[s3]',
-    'scikit-learn', 'scikit-image', 'tqdm', 'cartopy']  
+    'scikit-learn', 'scikit-image', 'tqdm']#, 'cartopy']  
 setup_keywords['zip_safe'] = False
 setup_keywords['use_2to3'] = False
 setup_keywords['packages'] = find_packages()
