+++ ulmo/llc/extract.py
@@ -2,7 +2,6 @@
 
 import os
 import numpy as np
-import datetime
 
 import pandas
 
@@ -10,12 +9,6 @@ import xarray as xr
 import h5py
 
 from sklearn.utils import shuffle
-
-from matplotlib import pyplot as plt
-
-import cartopy.crs as ccrs
-from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
-
 
 from functools import partial
 from concurrent.futures import ProcessPoolExecutor
@@ -30,9 +23,6 @@ from ulmo.llc import io as llc_io
 from ulmo import io as ulmo_io
 
 
-# Astronomy tools
-import astropy_healpix
-from astropy import units
 
 from IPython import embed
 
@@ -112,50 +102,6 @@ def build_CC_mask(filename=None, temp_bounds=(-3, 34),
     # Return
     return CC_mask
 
-
-def plot_extraction(llc_table, resol=None, cbar=False, s=0.01):
-
-    fig = plt.figure(figsize=(7, 4))
-    plt.clf()
-
-    tformM = ccrs.Mollweide()
-    tformP = ccrs.PlateCarree()
-
-    ax = plt.axes(projection=tformM)
-
-
-    cm = plt.get_cmap('Blues')
-    # Cut
-    #good = np.invert(hp_events.mask)
-    img = plt.scatter(x=llc_table.lon,
-        y=llc_table.lat,
-        s=s,
-        transform=tformP)
-
-    # Healpix?
-    if resol is not None:
-        nside = astropy_healpix.pixel_resolution_to_nside(0.5*units.deg)
-        hp = astropy_healpix.HEALPix(nside=nside)
-        hp_lon, hp_lat = hp.healpix_to_lonlat(np.arange(hp.npix))
-        img = plt.scatter(x=hp_lon.to('deg').value,
-            y=hp_lat.to('deg').value,
-            s=s,
-            color='r',
-            transform=tformP)
-
-    #
-    # Colorbar
-    if cbar:
-        cb = plt.colorbar(img, orientation='horizontal', pad=0.)
-        cb.ax.tick_params(labelsize=17)
-
-    # Coast lines
-    ax.coastlines(zorder=10)
-    ax.set_global()
-
-    plt.show()
-
-    return
 
 def preproc_image(item, pdict):
     """
